feat: stub dao 생성, service 테스트 추가
