+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,74 +1,13 @@
 package roomescape.reservation.dao;
 
-import java.sql.PreparedStatement;
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
 import roomescape.reservation.Reservation;
-import roomescape.reservationtime.ReservationTime;
 
-@Repository
-public class ReservationDao {
+public interface ReservationDao {
+    List<Reservation> findAll();
 
-    private JdbcTemplate jdbcTemplate;
+    Long create(Reservation reservation);
 
-    @Autowired
-    public ReservationDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Reservation> findAll() {
-        String sql = "SELECT \n"
-                + "    r.id as reservation_id, \n"
-                + "    r.name, \n"
-                + "    r.date, \n"
-                + "    t.id as time_id, \n"
-                + "    t.start_at as time_value \n"
-                + "FROM reservation as r \n"
-                + "inner join reservation_time as t \n"
-                + "on r.time_id = t.id\n";
-
-        return this.jdbcTemplate.query(sql,
-                (resultSet, rowNum) -> {
-                    ReservationTime reservationTime = new ReservationTime(
-                            resultSet.getLong("time_id"),
-                            resultSet.getObject("time_value", LocalTime.class)
-                    );
-
-                    return new Reservation(
-                            resultSet.getLong("reservation_id"),
-                            resultSet.getString("name"),
-                            resultSet.getObject("date", LocalDate.class),
-                            reservationTime
-                    );
-                });
-    }
-
-    public Long create(Reservation reservation) {
-        String sql = "insert into reservation (name, date, time_id) values (?, ?, ?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        this.jdbcTemplate.update(con -> {
-            PreparedStatement ps = con.prepareStatement(
-                    sql,
-                    new String[]{"id"}
-            );
-            ps.setString(1, reservation.getName());
-            ps.setString(2, reservation.getDate().toString());
-            ps.setLong(3, reservation.getReservationTime().getId());
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-
-    public int delete(Long id) {
-        String sql = "delete from reservation where id = ?";
-        return this.jdbcTemplate.update(sql, id);
-    }
+    int delete(Long id);
 }
+
