+++ src/main/java/roomescape/reservationtime/dao/JdbcReservationTimeDao.java
@@ -1,0 +1,71 @@
+package roomescape.reservationtime.dao;
+
+import java.sql.PreparedStatement;
+import java.time.LocalTime;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import roomescape.reservationtime.ReservationTime;
+
+@Repository
+public class JdbcReservationTimeDao implements ReservationTimeDao {
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "select * from reservation_time";
+        return this.jdbcTemplate.query(sql,
+                (resultSet, rowNum) -> {
+                    String timeString = resultSet.getString("start_at");
+
+                    return new ReservationTime(
+                            resultSet.getLong("id"),
+                            LocalTime.parse(timeString)
+                    );
+                });
+    }
+
+    @Override
+    public Long create(ReservationTime reservationTime) {
+        String sql = "insert into reservation_time (start_at) values (?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        this.jdbcTemplate.update(con -> {
+            PreparedStatement ps = con.prepareStatement(
+                    sql,
+                    new String[]{"id"}
+            );
+            LocalTime startAt = reservationTime.getStartAt();
+            ps.setString(1, startAt.toString());
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    @Override
+    public int delete(Long id) {
+        String sql = "delete from reservation_time where id = ?";
+        return this.jdbcTemplate.update(sql, id);
+    }
+
+    @Override
+    public ReservationTime findById(Long id) {
+        String sql = "select * from reservation_time where id = ?";
+
+        return this.jdbcTemplate.queryForObject(sql,
+                (resultSet, rowNum) -> {
+                    ReservationTime reservationTime = new ReservationTime(
+                            resultSet.getLong("id"),
+                            resultSet.getObject("start_at", LocalTime.class)
+                    );
+                    return reservationTime;
+                }, id
+        );
+    }
+}
