+++ src/test/java/roomescape/SpringJdbcTest.java
@@ -2,18 +2,14 @@ package roomescape;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
+import java.lang.reflect.Field;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.controller.ReservationController;
 
 import org.junit.jupiter.api.Test;
 
@@ -34,44 +30,21 @@ public class SpringJdbcTest {
         }
     }
 
-    @Test
-    void 데이터_조회_테스트() {
-        jdbcTemplate.update("INSERT INTO reservation (name, date, time) VALUES (?, ?, ?)", "포라", "2025-08-05", "15:40");
-
-        List<ReservationResponse> reservations = RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200).extract()
-                .jsonPath().getList(".", ReservationResponse.class);
-
-        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-
-        assertThat(reservations.size()).isEqualTo(count);
-    }
+    @Autowired
+    private ReservationController reservationController;
 
     @Test
-    void 데이터_추가_삭제_테스트() {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "포라");
-        params.put("date", "2025-08-05");
-        params.put("time", "10:00");
+    void JdbcTemplate_필드_삭제_확인_테스트() {
+        boolean isJdbcTemplateInjected = false;
 
-        RestAssured.given().log().all()
-                .contentType(ContentType.JSON)
-                .body(params)
-                .when().post("/reservations")
-                .then().log().all()
-                .statusCode(200);
+        for (Field field : reservationController.getClass().getDeclaredFields()) {
+            if (field.getType().equals(JdbcTemplate.class)) {
+                isJdbcTemplateInjected = true;
+                break;
+            }
+        }
 
-        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-        assertThat(count).isEqualTo(1);
-
-        RestAssured.given().log().all()
-                .when().delete("/reservations/1")
-                .then().log().all()
-                .statusCode(200);
-
-        Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
-        assertThat(countAfterDelete).isEqualTo(0);
+        assertThat(isJdbcTemplateInjected).isFalse();
     }
+
 }
