+++ src/test/java/roomescape/reservationtime/dao/JdbcReservationTimeDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.reservationtime.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -8,13 +8,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.reservationtime.ReservationTime;
-import roomescape.reservationtime.dao.ReservationTimeDao;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-@SpringBootTest
-class ReservationTimeDaoTest {
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+class JdbcReservationTimeDaoTest {
 
     @Autowired
     private ReservationTimeDao reservationTimeDao;
